requirements.txt
# requirements.txt - Reflecting original constraints before updates

langchain>=0.1.0,<0.4.0
langchain-community>=0.0.1,<0.4.0
langchain-huggingface>=0.0.1,<0.2.0
nltk>=3.8.1
pandas>=2.0.0,<3.0.0
requests>=2.31.0
gradio==5.23.2         # Note: Recommend latest 4.x like 4.31.0 if available within this range
chromadb>=0.4.17,<0.7.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.1
pyyaml>=6.0.1
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.23.2      # Added [standard] for common extras
datasets>=2.14.0               # Hugging Face datasets library
huggingface_hub>=0.19.0
pyarrow                       # No specific version constraint provided initially
sentencepiece                 # No specific version constraint provided initially
google-generativeai           # No specific version constraint provided initially
langchain-google-genai>=0.0.9

# Note: Pip will install the latest versions compatible with these constraints.
# Using this list might avoid some dependency conflicts compared to using all absolute latest versions.